refactor(jvm): split adapter seams for detection and scanning
