+++ internal/lang/jvm/adapter.go
@@ -2,20 +2,10 @@ package jvm
 
 import (
 	"context"
-	"encoding/xml"
-	"errors"
-	"fmt"
-	"io/fs"
-	"os"
-	"path/filepath"
-	"regexp"
-	"sort"
-	"strings"
 
 	"github.com/ben-ranford/lopper/internal/lang/shared"
 	"github.com/ben-ranford/lopper/internal/language"
 	"github.com/ben-ranford/lopper/internal/report"
-	"github.com/ben-ranford/lopper/internal/safeio"
 	"github.com/ben-ranford/lopper/internal/workspace"
 )
 
@@ -42,113 +32,6 @@ func NewAdapter() *Adapter {
 	adapter := &Adapter{}
 	adapter.AdapterLifecycle = language.NewAdapterLifecycle("jvm", []string{"java", "kotlin"}, adapter.DetectWithConfidence)
 	return adapter
-}
-
-func (a *Adapter) DetectWithConfidence(ctx context.Context, repoPath string) (language.Detection, error) {
-	_ = ctx
-	repoPath = shared.DefaultRepoPath(repoPath)
-
-	detection := language.Detection{}
-	roots := make(map[string]struct{})
-	if err := applyJVMRootSignals(repoPath, &detection, roots); err != nil {
-		return language.Detection{}, err
-	}
-
-	const maxFiles = 1024
-	visited := 0
-	err := filepath.WalkDir(repoPath, func(path string, entry fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-		return walkJVMDetectionEntry(path, entry, roots, &detection, &visited, maxFiles)
-	})
-	if err != nil && !errors.Is(err, fs.SkipAll) {
-		return language.Detection{}, err
-	}
-
-	return shared.FinalizeDetection(repoPath, detection, roots), nil
-}
-
-func walkJVMDetectionEntry(path string, entry fs.DirEntry, roots map[string]struct{}, detection *language.Detection, visited *int, maxFiles int) error {
-	if entry.IsDir() {
-		if shouldSkipDir(entry.Name()) {
-			return filepath.SkipDir
-		}
-		return nil
-	}
-	(*visited)++
-	if *visited > maxFiles {
-		return fs.SkipAll
-	}
-	updateJVMDetection(path, entry, roots, detection)
-	return nil
-}
-
-func applyJVMRootSignals(repoPath string, detection *language.Detection, roots map[string]struct{}) error {
-	rootSignals := []struct {
-		name       string
-		confidence int
-	}{
-		{name: pomXMLName, confidence: 55},
-		{name: buildGradleName, confidence: 45},
-		{name: buildGradleKTSName, confidence: 45},
-	}
-	for _, signal := range rootSignals {
-		path := filepath.Join(repoPath, signal.name)
-		if _, err := os.Stat(path); err == nil {
-			detection.Matched = true
-			detection.Confidence += signal.confidence
-			roots[repoPath] = struct{}{}
-		} else if !os.IsNotExist(err) {
-			return err
-		}
-	}
-	return nil
-}
-
-func updateJVMDetection(path string, entry fs.DirEntry, roots map[string]struct{}, detection *language.Detection) {
-	switch strings.ToLower(entry.Name()) {
-	case pomXMLName, buildGradleName, buildGradleKTSName:
-		detection.Matched = true
-		detection.Confidence += 10
-		roots[filepath.Dir(path)] = struct{}{}
-	}
-	switch strings.ToLower(filepath.Ext(path)) {
-	case ".java", ".kt", ".kts":
-		detection.Matched = true
-		detection.Confidence += 2
-		if root := sourceLayoutModuleRoot(path); root != "" {
-			roots[root] = struct{}{}
-		}
-	}
-}
-
-func sourceLayoutModuleRoot(path string) string {
-	normalized := filepath.ToSlash(filepath.Clean(path))
-	if normalized == "" {
-		return ""
-	}
-
-	segments := strings.Split(normalized, "/")
-	lastSrcIndex := -1
-	for index := 0; index+2 < len(segments); index++ {
-		if segments[index] != "src" {
-			continue
-		}
-		switch segments[index+2] {
-		case "java", "kotlin":
-			lastSrcIndex = index
-		}
-	}
-	if lastSrcIndex < 1 {
-		return ""
-	}
-
-	root := strings.Join(segments[:lastSrcIndex], "/")
-	if root == "" {
-		return ""
-	}
-	return filepath.FromSlash(root)
 }
 
 func (a *Adapter) Analyse(ctx context.Context, req language.Request) (report.Report, error) {
@@ -200,242 +83,6 @@ func resolveRemovalCandidateWeights(value *report.RemovalCandidateWeights) repor
 		return report.DefaultRemovalCandidateWeights()
 	}
 	return report.NormalizeRemovalCandidateWeights(*value)
-}
-
-type importBinding = shared.ImportRecord
-
-type fileScan struct {
-	Path    string
-	Package string
-	Imports []importBinding
-	Usage   map[string]int
-}
-
-type scanResult struct {
-	Files    []fileScan
-	Warnings []string
-}
-
-func scanRepo(ctx context.Context, repoPath string, depPrefixes map[string]string, depAliases map[string]string) (scanResult, error) {
-	result := scanResult{}
-	if repoPath == "" {
-		return result, fs.ErrInvalid
-	}
-
-	err := filepath.WalkDir(repoPath, func(path string, entry fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-		if ctx.Err() != nil {
-			return ctx.Err()
-		}
-		if entry.IsDir() {
-			if shouldSkipDir(entry.Name()) {
-				return filepath.SkipDir
-			}
-			return nil
-		}
-		return scanJVMSourceFile(repoPath, path, depPrefixes, depAliases, &result)
-	})
-	if err != nil {
-		return result, err
-	}
-
-	if len(result.Files) == 0 {
-		result.Warnings = append(result.Warnings, "no Java/Kotlin source files found for analysis")
-	}
-	return result, nil
-}
-
-func scanJVMSourceFile(repoPath string, path string, depPrefixes map[string]string, depAliases map[string]string, result *scanResult) error {
-	if !isSourceFile(path) {
-		return nil
-	}
-	var (
-		content []byte
-		err     error
-	)
-	if strings.TrimSpace(repoPath) == "" {
-		content, err = safeio.ReadFile(path)
-	} else {
-		content, err = safeio.ReadFileUnder(repoPath, path)
-	}
-	if err != nil {
-		return err
-	}
-	relativePath, err := filepath.Rel(repoPath, path)
-	if err != nil {
-		relativePath = path
-	}
-
-	filePackage := parsePackage(content)
-	imports := parseImports(content, relativePath, filePackage, depPrefixes, depAliases)
-	result.Files = append(result.Files, fileScan{
-		Path:    relativePath,
-		Package: filePackage,
-		Imports: imports,
-		Usage:   countUsage(content, imports),
-	})
-	return nil
-}
-
-func isSourceFile(path string) bool {
-	switch strings.ToLower(filepath.Ext(path)) {
-	case ".java", ".kt", ".kts":
-		return true
-	default:
-		return false
-	}
-}
-
-var (
-	packagePattern          = regexp.MustCompile(`(?m)^\s*package\s+([A-Za-z_][A-Za-z0-9_\.]*)\s*;?\s*$`)
-	importPattern           = regexp.MustCompile(`(?m)^\s*import\s+(?:static\s+)?([A-Za-z_][A-Za-z0-9_\.]*)(\.\*)?(?:\s+as\s+([A-Za-z_][A-Za-z0-9_]*))?\s*;?\s*$`)
-	pomPropertyTokenPattern = regexp.MustCompile(`\$\{([^}]+)\}`)
-)
-
-const importPatternMatchGroups = 4
-
-func parsePackage(content []byte) string {
-	matches := packagePattern.FindSubmatch(content)
-	if len(matches) != 2 {
-		return ""
-	}
-	return strings.TrimSpace(string(matches[1]))
-}
-
-func parseImports(content []byte, filePath string, filePackage string, depPrefixes map[string]string, depAliases map[string]string) []importBinding {
-	return shared.ParseImportLines(content, filePath, func(line string, _ int) []shared.ImportRecord {
-		line = stripLineComment(line)
-		matches := importPattern.FindStringSubmatch(line)
-		if len(matches) != importPatternMatchGroups {
-			return nil
-		}
-		module := strings.TrimSpace(matches[1])
-		if module == "" || shouldIgnoreImport(module, filePackage) {
-			return nil
-		}
-
-		dependency := resolveDependency(module, depPrefixes, depAliases)
-		if dependency == "" {
-			dependency = fallbackDependency(module)
-		}
-		if dependency == "" {
-			return nil
-		}
-
-		record, ok := buildImportRecord(matches, module, dependency)
-		if !ok {
-			return nil
-		}
-
-		return []shared.ImportRecord{record}
-	})
-}
-
-func buildImportRecord(matches []string, module string, dependency string) (shared.ImportRecord, bool) {
-	wildcard := strings.TrimSpace(matches[2]) == ".*"
-	symbol := lastModuleSegment(module)
-	if wildcard {
-		symbol = "*"
-	}
-	if symbol == "" {
-		return shared.ImportRecord{}, false
-	}
-
-	localName := symbol
-	if alias := strings.TrimSpace(matches[3]); alias != "" && !wildcard {
-		localName = alias
-	}
-
-	return shared.ImportRecord{
-		Dependency: dependency,
-		Module:     module,
-		Name:       symbol,
-		Local:      localName,
-		Wildcard:   wildcard,
-	}, true
-}
-
-func stripLineComment(line string) string {
-	return shared.StripLineComment(line, "//")
-}
-
-func shouldIgnoreImport(module, filePackage string) bool {
-	module = strings.TrimSpace(module)
-	if module == "" {
-		return true
-	}
-
-	stdlibPrefixes := []string{
-		"java.", "javax.", "kotlin.", "jdk.", "sun.",
-	}
-	for _, prefix := range stdlibPrefixes {
-		if strings.HasPrefix(module, prefix) {
-			return true
-		}
-	}
-
-	if filePackage != "" {
-		if module == filePackage || strings.HasPrefix(module, filePackage+".") {
-			return true
-		}
-	}
-	return false
-}
-
-func resolveDependency(module string, depPrefixes map[string]string, depAliases map[string]string) string {
-	best := ""
-	bestLen := 0
-
-	for prefix, dependency := range depPrefixes {
-		if module == prefix || strings.HasPrefix(module, prefix+".") {
-			if len(prefix) > bestLen {
-				best = dependency
-				bestLen = len(prefix)
-			}
-		}
-	}
-	if best != "" {
-		return best
-	}
-
-	parts := strings.Split(module, ".")
-	for i := len(parts); i >= 1; i-- {
-		key := strings.Join(parts[:i], ".")
-		if dependency, ok := depAliases[key]; ok {
-			return dependency
-		}
-	}
-
-	return ""
-}
-
-func fallbackDependency(module string) string {
-	parts := strings.Split(module, ".")
-	if len(parts) >= 2 {
-		return normalizeDependencyID(parts[0] + "." + parts[1])
-	}
-	if len(parts) == 1 {
-		return normalizeDependencyID(parts[0])
-	}
-	return ""
-}
-
-func lastModuleSegment(module string) string {
-	parts := strings.Split(module, ".")
-	if len(parts) == 0 {
-		return ""
-	}
-	return strings.TrimSpace(parts[len(parts)-1])
-}
-
-func firstContentColumn(line string) int {
-	return shared.FirstContentColumn(line)
-}
-
-func countUsage(content []byte, imports []importBinding) map[string]int {
-	return shared.CountUsage(content, imports)
 }
 
 func buildDependencyReport(dependency string, scan scanResult) (report.DependencyReport, []string) {
@@ -495,507 +142,4 @@ func normalizeDependencyID(value string) string {
 
 func shouldSkipDir(name string) bool {
 	return shared.ShouldSkipDir(name, jvmSkippedDirectories)
-}
-
-type dependencyDescriptor struct {
-	Name     string
-	Group    string
-	Artifact string
-}
-
-type pomProjectModel struct {
-	GroupID              string               `xml:"groupId"`
-	ArtifactID           string               `xml:"artifactId"`
-	Version              string               `xml:"version"`
-	Parent               pomParentModel       `xml:"parent"`
-	Properties           pomPropertiesModel   `xml:"properties"`
-	Dependencies         []pomDependencyModel `xml:"dependencies>dependency"`
-	DependencyManagement struct {
-		Dependencies []pomDependencyModel `xml:"dependencies>dependency"`
-	} `xml:"dependencyManagement"`
-}
-
-type pomParentModel struct {
-	GroupID string `xml:"groupId"`
-	Version string `xml:"version"`
-}
-
-type pomPropertiesModel struct {
-	Properties []pomPropertyModel `xml:",any"`
-}
-
-type pomPropertyModel struct {
-	XMLName xml.Name
-	Value   string `xml:",chardata"`
-}
-
-type pomDependencyModel struct {
-	GroupID    string `xml:"groupId"`
-	ArtifactID string `xml:"artifactId"`
-	Version    string `xml:"version"`
-	Type       string `xml:"type"`
-	Scope      string `xml:"scope"`
-}
-
-type pomDependencyKind int
-
-const (
-	pomDependencyDirect pomDependencyKind = iota
-	pomDependencyManaged
-)
-
-func collectDeclaredDependencies(repoPath string) ([]dependencyDescriptor, map[string]string, map[string]string, []string) {
-	descriptors := make([]dependencyDescriptor, 0)
-	warnings := make([]string, 0)
-
-	pomDescriptors, pomWarnings := parsePomDependenciesWithWarnings(repoPath)
-	gradleDescriptors, gradleWarnings := parseGradleDependenciesWithWarnings(repoPath)
-	descriptors = append(descriptors, pomDescriptors...)
-	descriptors = append(descriptors, gradleDescriptors...)
-	warnings = append(warnings, pomWarnings...)
-	warnings = append(warnings, gradleWarnings...)
-
-	descriptors = dedupeAndSortDescriptors(descriptors)
-	prefixes, aliases := buildDescriptorLookups(descriptors)
-	return descriptors, prefixes, aliases, warnings
-}
-
-func dedupeAndSortDescriptors(descriptors []dependencyDescriptor) []dependencyDescriptor {
-	unique := make(map[string]dependencyDescriptor)
-	for _, descriptor := range descriptors {
-		key := descriptor.Group + ":" + descriptor.Artifact
-		if descriptor.Group == "" {
-			key = descriptor.Name
-		}
-		unique[key] = descriptor
-	}
-	items := make([]dependencyDescriptor, 0, len(unique))
-	for _, descriptor := range unique {
-		items = append(items, descriptor)
-	}
-	sort.Slice(items, func(i, j int) bool {
-		if items[i].Name == items[j].Name {
-			return items[i].Group < items[j].Group
-		}
-		return items[i].Name < items[j].Name
-	})
-	return items
-}
-
-func buildDescriptorLookups(descriptors []dependencyDescriptor) (map[string]string, map[string]string) {
-	prefixes := make(map[string]string)
-	aliases := make(map[string]string)
-	for _, descriptor := range descriptors {
-		name := normalizeDependencyID(descriptor.Name)
-		addGroupLookups(prefixes, aliases, name, descriptor.Group)
-		addArtifactLookups(prefixes, aliases, name, descriptor.Group, descriptor.Artifact)
-	}
-	return prefixes, aliases
-}
-
-type lookupKeyStrategy func(group string, artifact string) ([]string, []string)
-
-func addGroupLookups(prefixes map[string]string, aliases map[string]string, name string, group string) {
-	addLookupByStrategy(prefixes, aliases, name, group, "", groupLookupStrategy)
-}
-
-func addArtifactLookups(prefixes map[string]string, aliases map[string]string, name string, group string, artifact string) {
-	addLookupByStrategy(prefixes, aliases, name, group, artifact, artifactLookupStrategy)
-}
-
-func addLookupByStrategy(prefixes map[string]string, aliases map[string]string, name string, group string, artifact string, strategy lookupKeyStrategy) {
-	prefixKeys, aliasKeys := strategy(group, artifact)
-	for _, key := range prefixKeys {
-		prefixes[key] = name
-	}
-	for _, key := range aliasKeys {
-		aliases[key] = name
-	}
-}
-
-func groupLookupStrategy(group, _ string) ([]string, []string) {
-	if group == "" {
-		return nil, nil
-	}
-	group = strings.TrimSpace(group)
-	prefixes := []string{group}
-	aliases := []string{group}
-	parts := strings.Split(group, ".")
-	if len(parts) >= 2 {
-		aliases = append(aliases, parts[0]+"."+parts[1], parts[len(parts)-1])
-	}
-	return prefixes, aliases
-}
-
-func artifactLookupStrategy(group, artifact string) ([]string, []string) {
-	if artifact == "" {
-		return nil, nil
-	}
-	artifact = strings.ReplaceAll(strings.TrimSpace(artifact), "-", ".")
-	prefixes := make([]string, 0, 1)
-	aliases := make([]string, 0, 1)
-	if group != "" && artifact != "" {
-		prefixes = append(prefixes, group+"."+artifact)
-	}
-	if artifact != "" {
-		aliases = append(aliases, artifact)
-	}
-	return prefixes, aliases
-}
-
-func parsePomDependencies(repoPath string) []dependencyDescriptor {
-	descriptors, _ := parsePomDependenciesWithWarnings(repoPath)
-	return descriptors
-}
-
-func parsePomDependenciesWithWarnings(repoPath string) ([]dependencyDescriptor, []string) {
-	pomParser := func(path, content string) ([]dependencyDescriptor, []string) {
-		return parsePomDependencyContent(relativeBuildFilePath(repoPath, path), content)
-	}
-	return parseBuildFilesWithWarnings(repoPath, pomParser, pomXMLName)
-}
-
-func parsePomDependencyContent(relativePath, content string) ([]dependencyDescriptor, []string) {
-	var project pomProjectModel
-	if err := xml.Unmarshal([]byte(content), &project); err != nil {
-		return nil, []string{fmt.Sprintf("unable to parse Maven POM %s: %v", relativePath, err)}
-	}
-
-	propertyMap := buildPomPropertyMap(project)
-	directDescriptors, directWarnings := parsePomDependencyList(project.Dependencies, propertyMap, pomDependencyDirect, relativePath)
-	managedDescriptors, managedWarnings := parsePomDependencyList(project.DependencyManagement.Dependencies, propertyMap, pomDependencyManaged, relativePath)
-
-	descriptors := make([]dependencyDescriptor, 0, len(directDescriptors)+len(managedDescriptors))
-	descriptors = append(descriptors, directDescriptors...)
-	descriptors = append(descriptors, managedDescriptors...)
-
-	warnings := make([]string, 0, len(directWarnings)+len(managedWarnings))
-	warnings = append(warnings, directWarnings...)
-	warnings = append(warnings, managedWarnings...)
-
-	return dedupeAndSortDescriptors(descriptors), shared.DedupeWarnings(warnings)
-}
-
-func parsePomDependencyList(dependencies []pomDependencyModel, propertyMap map[string]string, kind pomDependencyKind, relativePath string) ([]dependencyDescriptor, []string) {
-	descriptors := make([]dependencyDescriptor, 0, len(dependencies))
-	warnings := make([]string, 0)
-	for _, dependency := range dependencies {
-		descriptor, warning := parsePomDependency(dependency, propertyMap, kind, relativePath)
-		if descriptor.Group != "" && descriptor.Artifact != "" {
-			descriptors = append(descriptors, descriptor)
-		}
-		if warning != "" {
-			warnings = append(warnings, warning)
-		}
-	}
-	return descriptors, warnings
-}
-
-func parsePomDependency(dependency pomDependencyModel, propertyMap map[string]string, kind pomDependencyKind, relativePath string) (dependencyDescriptor, string) {
-	group, unresolvedGroup := resolvePomPropertyValue(dependency.GroupID, propertyMap)
-	artifact, unresolvedArtifact := resolvePomPropertyValue(dependency.ArtifactID, propertyMap)
-	if unresolvedGroup || unresolvedArtifact || group == "" || artifact == "" {
-		return dependencyDescriptor{}, ""
-	}
-
-	descriptor := dependencyDescriptor{
-		Name:     artifact,
-		Group:    group,
-		Artifact: artifact,
-	}
-	if kind != pomDependencyManaged {
-		return descriptor, ""
-	}
-
-	version, unresolvedVersion := resolvePomPropertyValue(dependency.Version, propertyMap)
-	if !isPomImportedBOM(dependency) {
-		if version == "" || unresolvedVersion {
-			return descriptor, fmt.Sprintf("unable to resolve managed Maven version for %s:%s in %s", group, artifact, relativePath)
-		}
-		return descriptor, ""
-	}
-
-	if version == "" || unresolvedVersion {
-		return descriptor, fmt.Sprintf("unable to resolve imported Maven BOM version for %s:%s in %s", group, artifact, relativePath)
-	}
-	return descriptor, ""
-}
-
-func isPomImportedBOM(dependency pomDependencyModel) bool {
-	return strings.EqualFold(strings.TrimSpace(dependency.Type), "pom") &&
-		strings.EqualFold(strings.TrimSpace(dependency.Scope), "import")
-}
-
-func buildPomPropertyMap(project pomProjectModel) map[string]string {
-	properties := make(map[string]string)
-	for _, property := range project.Properties.Properties {
-		key := strings.TrimSpace(property.XMLName.Local)
-		value := strings.TrimSpace(property.Value)
-		if key == "" || value == "" {
-			continue
-		}
-		properties[key] = value
-	}
-
-	groupID := strings.TrimSpace(project.GroupID)
-	if groupID == "" {
-		groupID = strings.TrimSpace(project.Parent.GroupID)
-	}
-	version := strings.TrimSpace(project.Version)
-	if version == "" {
-		version = strings.TrimSpace(project.Parent.Version)
-	}
-	artifactID := strings.TrimSpace(project.ArtifactID)
-
-	setPomPropertyValue(properties, "project.groupId", groupID)
-	setPomPropertyValue(properties, "pom.groupId", groupID)
-	setPomPropertyValue(properties, "groupId", groupID)
-
-	setPomPropertyValue(properties, "project.version", version)
-	setPomPropertyValue(properties, "pom.version", version)
-	setPomPropertyValue(properties, "version", version)
-
-	setPomPropertyValue(properties, "project.artifactId", artifactID)
-	setPomPropertyValue(properties, "pom.artifactId", artifactID)
-	setPomPropertyValue(properties, "artifactId", artifactID)
-
-	setPomPropertyValue(properties, "project.parent.groupId", strings.TrimSpace(project.Parent.GroupID))
-	setPomPropertyValue(properties, "project.parent.version", strings.TrimSpace(project.Parent.Version))
-	return properties
-}
-
-func setPomPropertyValue(properties map[string]string, key, value string) {
-	key = strings.TrimSpace(key)
-	value = strings.TrimSpace(value)
-	if key == "" || value == "" {
-		return
-	}
-	properties[key] = value
-}
-
-func resolvePomPropertyValue(value string, properties map[string]string) (string, bool) {
-	value = strings.TrimSpace(value)
-	if value == "" {
-		return "", false
-	}
-	unresolved := false
-	for iteration := 0; iteration < 8; iteration++ {
-		updated, replaced, missing := replacePomPropertyTokens(value, properties)
-		unresolved = unresolved || missing
-		if !replaced {
-			break
-		}
-		value = updated
-	}
-	if pomPropertyTokenPattern.MatchString(value) {
-		unresolved = true
-	}
-	return strings.TrimSpace(value), unresolved
-}
-
-func replacePomPropertyTokens(value string, properties map[string]string) (string, bool, bool) {
-	matches := pomPropertyTokenPattern.FindAllStringSubmatch(value, -1)
-	if len(matches) == 0 {
-		return value, false, false
-	}
-
-	updated := value
-	replaced := false
-	unresolved := false
-	for _, match := range matches {
-		token, replacement, ok := pomPropertyReplacement(match, properties)
-		if !ok {
-			unresolved = unresolved || len(match) == 2
-			continue
-		}
-		updated = strings.ReplaceAll(updated, token, replacement)
-		replaced = true
-	}
-	return updated, replaced, unresolved
-}
-
-func pomPropertyReplacement(match []string, properties map[string]string) (string, string, bool) {
-	if len(match) != 2 {
-		return "", "", false
-	}
-
-	key := strings.TrimSpace(match[1])
-	replacement, ok := properties[key]
-	if !ok {
-		return match[0], "", false
-	}
-
-	replacement = strings.TrimSpace(replacement)
-	if replacement == "" {
-		return match[0], "", false
-	}
-	return match[0], replacement, true
-}
-
-func parseGradleDependencies(repoPath string) []dependencyDescriptor {
-	descriptors, _ := parseGradleDependenciesWithWarnings(repoPath)
-	return descriptors
-}
-
-func parseGradleDependenciesWithWarnings(repoPath string) ([]dependencyDescriptor, []string) {
-	pattern := regexp.MustCompile(`(?m)(?:implementation|api|compileOnly|runtimeOnly|testImplementation|testRuntimeOnly|kapt)\s*\(?\s*["']([^:"'\s]+):([^:"'\s]+):[^"'\s]+["']\s*\)?`)
-	catalogResolver, warnings := shared.LoadGradleCatalogResolver(repoPath)
-	gradleParser := func(path, content string) ([]dependencyDescriptor, []string) {
-		descriptors := parseGradleMatches(content, pattern)
-		catalogDescriptors, catalogWarnings := catalogResolver.ParseDependencyReferences(path, content)
-		for _, descriptor := range catalogDescriptors {
-			descriptors = append(descriptors, dependencyDescriptor{
-				Name:     descriptor.Artifact,
-				Group:    descriptor.Group,
-				Artifact: descriptor.Artifact,
-			})
-		}
-		return dedupeAndSortDescriptors(descriptors), catalogWarnings
-	}
-	descriptors, parseWarnings := parseBuildFilesWithWarnings(repoPath, gradleParser, buildGradleName, buildGradleKTSName)
-	warnings = append(warnings, parseWarnings...)
-	return descriptors, shared.DedupeWarnings(warnings)
-}
-
-func parseGradleMatches(content string, pattern *regexp.Regexp) []dependencyDescriptor {
-	return parseDependencyDescriptorsFromMatches(pattern.FindAllStringSubmatch(content, -1))
-}
-
-func parseDependencyDescriptorsFromMatches(matches [][]string) []dependencyDescriptor {
-	descriptors := make([]dependencyDescriptor, 0, len(matches))
-	for _, match := range matches {
-		if len(match) != 3 {
-			continue
-		}
-		group := strings.TrimSpace(match[1])
-		artifact := strings.TrimSpace(match[2])
-		if group == "" || artifact == "" {
-			continue
-		}
-		descriptors = append(descriptors, dependencyDescriptor{
-			Name:     artifact,
-			Group:    group,
-			Artifact: artifact,
-		})
-	}
-	return descriptors
-}
-
-func parseBuildFiles(repoPath string, primaryName string, parser func(content string) []dependencyDescriptor, additionalNames ...string) []dependencyDescriptor {
-	names := append([]string{primaryName}, additionalNames...)
-	descriptors := make([]dependencyDescriptor, 0)
-	seen := make(map[string]struct{})
-
-	err := filepath.WalkDir(repoPath, func(path string, entry fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-		return parseBuildFileEntry(repoPath, path, entry, names, parser, seen, &descriptors)
-	})
-	if err != nil {
-		return descriptors
-	}
-	return descriptors
-}
-
-func parseBuildFilesWithWarnings(repoPath string, parser func(path, content string) ([]dependencyDescriptor, []string), names ...string) ([]dependencyDescriptor, []string) {
-	collector := buildFileWarningCollector{
-		repoPath: repoPath,
-		parser:   parser,
-		names:    names,
-		seen:     make(map[string]struct{}),
-	}
-	err := filepath.WalkDir(repoPath, collector.visit)
-	if err != nil {
-		collector.warnings = append(collector.warnings, err.Error())
-	}
-	return collector.descriptors, shared.DedupeWarnings(collector.warnings)
-}
-
-func parseBuildFileEntry(repoPath string, path string, entry fs.DirEntry, names []string, parser func(content string) []dependencyDescriptor, seen map[string]struct{}, descriptors *[]dependencyDescriptor) error {
-	if entry.IsDir() {
-		if shouldSkipDir(entry.Name()) {
-			return filepath.SkipDir
-		}
-		return nil
-	}
-	fileName := strings.ToLower(entry.Name())
-	if !matchesBuildFile(fileName, names) {
-		return nil
-	}
-
-	content, err := safeio.ReadFileUnder(repoPath, path)
-	if err != nil {
-		return nil
-	}
-	for _, descriptor := range parser(string(content)) {
-		key := descriptor.Group + ":" + descriptor.Artifact
-		if _, ok := seen[key]; ok {
-			continue
-		}
-		seen[key] = struct{}{}
-		*descriptors = append(*descriptors, descriptor)
-	}
-	return nil
-}
-
-type buildFileWarningCollector struct {
-	repoPath    string
-	parser      func(path, content string) ([]dependencyDescriptor, []string)
-	names       []string
-	seen        map[string]struct{}
-	descriptors []dependencyDescriptor
-	warnings    []string
-}
-
-func (c *buildFileWarningCollector) visit(path string, entry fs.DirEntry, err error) error {
-	if err != nil {
-		return err
-	}
-	if entry.IsDir() {
-		if shouldSkipDir(entry.Name()) {
-			return filepath.SkipDir
-		}
-		return nil
-	}
-	if !matchesBuildFile(strings.ToLower(entry.Name()), c.names) {
-		return nil
-	}
-	content, readErr := safeio.ReadFileUnder(c.repoPath, path)
-	if readErr != nil {
-		c.warnings = append(c.warnings, formatBuildFileReadWarning(c.repoPath, path, readErr))
-		return nil
-	}
-	items, parseWarnings := c.parser(path, string(content))
-	c.warnings = append(c.warnings, parseWarnings...)
-	for _, descriptor := range items {
-		key := descriptor.Group + ":" + descriptor.Artifact
-		if _, ok := c.seen[key]; ok {
-			continue
-		}
-		c.seen[key] = struct{}{}
-		c.descriptors = append(c.descriptors, descriptor)
-	}
-	return nil
-}
-
-func formatBuildFileReadWarning(repoPath, path string, err error) string {
-	return "unable to read " + relativeBuildFilePath(repoPath, path) + ": " + err.Error()
-}
-
-func relativeBuildFilePath(repoPath, path string) string {
-	relPath := path
-	if rel, relErr := filepath.Rel(repoPath, path); relErr == nil {
-		relPath = rel
-	}
-	return filepath.ToSlash(relPath)
-}
-
-func matchesBuildFile(fileName string, names []string) bool {
-	for _, name := range names {
-		if strings.EqualFold(fileName, name) {
-			return true
-		}
-	}
-	return false
 }
